Improve configuration mechanism
